+++ package.json
@@ -5,8 +5,14 @@
   "main": "index.js",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "cloudinary": "^1.14.0",
+    "connect-multiparty": "^2.2.0",
+    "datauri": "^2.0.0",
+    "esm": "^3.2.25",
+    "express": "^4.17.1",
     "express-session": "^1.16.2",
-    "mongoose": "^5.6.1"
+    "mongoose": "^5.6.1",
+    "multer": "^1.4.1"
   },
   "devDependencies": {},
   "scripts": {
